Increase amount of AlignedHits
As a temporary fix, we increase the number
of AlignedHits of the fingerprinter matching procedure
in order to reduce false positives. The default config
was 4 hits, which corresponds to about 1 sec
of audio at a resolution of 4fp/sec. We observed
a larger number of false positives at the range
of matches around 1 second ~ 4 hits. Therefore we increase
the amount of matches to 2 seconds, but in the long
term we need to fix this properly, e.g., including the
track length.